Split Popup into multiple components
